requirements.txt
fastapi
uvicorn
python-dotenv
line-bot-sdk
opencv-python
mediapipe
requests
